backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
email-validator==2.2.0
alembic==1.14.1
httpx==0.28.1
